requirements.txt
accelerate                0.22.0.dev0
aiohttp                   3.8.5
aiosignal                 1.3.1
altair                    5.0.1
async-timeout             4.0.2
attrs                     23.1.0
bitsandbytes              0.40.2
blinker                   1.6.2
cachetools                5.3.1
certifi                   2023.5.7
charset-normalizer        3.2.0
click                     8.1.6
cmake                     3.25.0
datasets                  2.13.1
decorator                 5.1.1
dill                      0.3.6
filelock                  3.9.0
frozenlist                1.4.0
fsspec                    2023.6.0
gitdb                     4.0.10
GitPython                 3.1.32
huggingface-hub           0.16.4
idna                      3.4
importlib-metadata        6.8.0
Jinja2                    3.1.2
jsonschema                4.18.4
jsonschema-specifications 2023.7.1
lit                       15.0.7
markdown-it-py            3.0.0
MarkupSafe                2.1.3
mdurl                     0.1.2
mpmath                    1.2.1
multidict                 6.0.4
multiprocess              0.70.14
networkx                  3.0
numpy                     1.25.1
packaging                 23.1
pandas                    2.0.3
peft                      0.5.0.dev0
Pillow                    9.5.0
pip                       22.0.2
protobuf                  4.23.4
psutil                    5.9.5
psycopg2-binary           2.9.9
pyarrow                   12.0.1
pydeck                    0.8.0
Pygments                  2.15.1
Pympler                   1.0.1
python-dateutil           2.8.2
python-dotenv             1.0.0
pytz                      2023.3
pytz-deprecation-shim     0.1.0.post0
PyYAML                    6.0.1
referencing               0.30.0
regex                     2023.6.3
requests                  2.31.0
rich                      13.4.2
rpds-py                   0.9.2
safetensors               0.3.1
scipy                     1.11.1
sentencepiece             0.1.99
setuptools                59.6.0
six                       1.16.0
smmap                     5.0.0
streamlit                 1.25.0
sympy                     1.11.1
tenacity                  8.2.2
tokenizers                0.13.3
toml                      0.10.2
toolz                     0.12.0
torch                     2.0.1+cu118
torchaudio                2.0.2+cu118
torchvision               0.15.2+cu118
tornado                   6.3.2
tqdm                      4.65.0
transformers              4.31.0
triton                    2.0.0
typing_extensions         4.7.1
tzdata                    2023.3
tzlocal                   4.3.1
urllib3                   2.0.4
validators                0.20.0
watchdog                  3.0.0
xxhash                    3.2.0
yarl                      1.9.2
zipp                      3.16.2